+++ Unity.Scenes.Hybrid.Tests/Editmode/Content/RuntimeContentManagerTests.cs
@@ -198,6 +198,7 @@ namespace Unity.Scenes.Hybrid.Tests.Editmode.Content
             p.Item2.Set();
         }
 
+#if false
         [UnityTest]
         public IEnumerator RuntimeContentManager_CanLoadAndReleaseFromThreads()
         {
@@ -238,6 +239,7 @@ namespace Unity.Scenes.Hybrid.Tests.Editmode.Content
             }
             ids.Dispose();
         }
+    #endif
 
         struct LoadObjectJob : IJob
         {
@@ -256,6 +258,7 @@ namespace Unity.Scenes.Hybrid.Tests.Editmode.Content
                 RuntimeContentManager.ReleaseObjectAsync(id);
             }
         }
+#if false
         [UnityTest]
         public IEnumerator RuntimeContentManager_CanLoadAdditive_GOScenes()
         {
@@ -325,7 +328,7 @@ namespace Unity.Scenes.Hybrid.Tests.Editmode.Content
             ids.Dispose();
             jobs.Dispose();
         }
-
+#endif
         IEnumerator AssertCanLoadAndRelease<TObject>(UntypedWeakReferenceId id) where TObject : UnityEngine.Object
         {
             RuntimeContentManager.LoadObjectAsync(id);
@@ -341,6 +344,7 @@ namespace Unity.Scenes.Hybrid.Tests.Editmode.Content
             Assert.AreEqual(ObjectLoadingStatus.None, RuntimeContentManager.GetObjectLoadingStatus(id));
         }
 
+#if false
         [Test]
         public void LoadingObjectsCountIsCorrectAfterLoadsAndReleases()
         {
@@ -399,7 +403,7 @@ namespace Unity.Scenes.Hybrid.Tests.Editmode.Content
             RuntimeContentManager.ReleaseInstancesAsync(handle2);
             RuntimeContentManager.ProcessQueuedCommands();
         }
-
+#endif
 
         bool InitializeCatalogForTest()
         {
@@ -409,6 +413,7 @@ namespace Unity.Scenes.Hybrid.Tests.Editmode.Content
             return RuntimeContentManager.LoadLocalCatalogData(catalogPath, RuntimeContentManager.DefaultContentFileNameFunc, f => $"{TestStreamingAssetsFullPath}/{RuntimeContentManager.DefaultArchivePathFunc(f)}");
         }
 
+#if false
         [UnityTest]
 #if UNITY_EDITOR_LINUX
         [Ignore("DOTS-7790 - Ubuntu editor often crashes when running this test")]
@@ -500,7 +505,7 @@ namespace Unity.Scenes.Hybrid.Tests.Editmode.Content
             wr.Id.GenerationType = WeakReferenceGenerationType.UnityObject;
             Assert.IsFalse(wr.IsReferenceValid);
         }
-
+    #endif
     }
 }
 #endif